+++ examples/hello_world.py
@@ -11,7 +11,7 @@ schema = build_schema(
 
 
 @schema.resolver("Query.hello")
-def resolve_hello(*_, value):
+def resolve_hello(*, value):
     return "Hello {}!".format(value)