+++ py_gql/__init__.py
@@ -8,8 +8,7 @@ query language for Python 3.5+.
 
 The main :mod:`py_gql` package provides the minimum required to build GraphQL
 schemas and execute queries against them while the relevant submodules allow
-you to customize the library's behaviour or implement your own GraphQL layer
-on top of :mod:`py_gql`.
+you implement custom behaviours and runtimes.
 """
 
 # flake8: noqa
@@ -18,7 +17,7 @@ from ._pkg import __version__  # isort:skip
 
 from . import lang, schema, tracers, utilities
 from ._graphql import graphql, graphql_blocking, process_graphql_query
-from .execution import GraphQLExtension, GraphQLResult, ResolveInfo
+from .execution import GraphQLResult, ResolveInfo
 from .sdl import build_schema
 
 __all__ = (
@@ -27,7 +26,6 @@ __all__ = (
     "graphql_blocking",
     "process_graphql_query",
     "GraphQLResult",
-    "GraphQLExtension",
     "ResolveInfo",
     "build_schema",
 )