+++ py_gql/execution/blocking_executor.py
@@ -20,7 +20,7 @@ class BlockingExecutor(Executor):
         This is aimed to be used internally to optimise the blocking execution
         case while keeping the base `Executor` class as generic as possible by
         side-stepping some of the operations that need to happen when working
-        with arbitrary wrapper types such as Awaitable, as a result this
+        with arbitrary wrapper types such as Awaitable. As a result this
         overrides much more of the base class than should be necessary to
         implement custom executors and should not be taken as an example.
     """