+++ py_gql/execution/runtime/__init__.py
@@ -7,10 +7,10 @@ from .base import Runtime, SubscriptionRuntime
 from .blocking import BlockingRuntime
 from .threadpool import ThreadPoolRuntime
 
-__all__ = (
-    "AsyncIORuntime",
-    "BlockingRuntime",
+__all__ = [
     "Runtime",
     "SubscriptionRuntime",
+    "BlockingRuntime",
+    "AsyncIORuntime",
     "ThreadPoolRuntime",
-)
+]